+++ cloudbuild.yaml
@@ -1,0 +1,86 @@
+# Modified from https://cloud.google.com/python/django/run
+
+# [START cloudrun_django_cloudbuild]
+steps:
+  - id: "build image"
+    name: "gcr.io/cloud-builders/docker"
+    # args:[]"${_REGION}-docker.pkg.dev/${PROJECT_ID}/${_ARTIFACT_REGISTRY_REPO}/${_SERVICE_NAME}:${COMMIT_SHA}",
+    args: ["build", "-t", "${_REGION}-docker.pkg.dev/${PROJECT_ID}/${_ARTIFACT_REGISTRY_REPO}/${_SERVICE_NAME}", "."]
+    # args: ["build", "-t", "gcr.io/${PROJECT_ID}/${_SERVICE_NAME}", "."]
+
+  - id: "push image"
+    name: "gcr.io/cloud-builders/docker"
+    # args: ["push", "gcr.io/${PROJECT_ID}/${_SERVICE_NAME}"]
+    args: ["push", "${_REGION}-docker.pkg.dev/${PROJECT_ID}/${_ARTIFACT_REGISTRY_REPO}/${_SERVICE_NAME}"]
+
+  - id: "apply migrations"
+    name: "gcr.io/google-appengine/exec-wrapper"
+    args:
+      [
+        "-i",
+        # "gcr.io/$PROJECT_ID/${_SERVICE_NAME}",
+        "${_REGION}-docker.pkg.dev/${PROJECT_ID}/${_ARTIFACT_REGISTRY_REPO}/${_SERVICE_NAME}",
+        "-s",
+        "${PROJECT_ID}:${_REGION}:${_INSTANCE_NAME}",
+        "-e",
+        "SETTINGS_NAME=${_SECRET_SETTINGS_NAME}",
+        "--",
+        "python",
+        "manage.py",
+        "migrate",
+      ]
+
+  # - id: "collect static"
+  #   name: "gcr.io/google-appengine/exec-wrapper"
+  #   args:
+  #     [
+  #       "-i",
+  #       "gcr.io/$PROJECT_ID/${_SERVICE_NAME}",
+  #       "-s",
+  #       "${PROJECT_ID}:${_REGION}:${_INSTANCE_NAME}",
+  #       "-e",
+  #       "SETTINGS_NAME=${_SECRET_SETTINGS_NAME}",
+  #       "--",
+  #       "python",
+  #       "manage.py",
+  #       "collectstatic",
+  #       "--verbosity",
+  #       "2",
+  #       "--no-input",
+  #     ]
+
+  - id: "deploy to cloud run"
+    name: "gcr.io/google.com/cloudsdktool/cloud-sdk"
+    entrypoint: gcloud
+    args:
+      [
+        "run",
+        "deploy",
+        "${_SERVICE_NAME}",
+        "--image",
+        # "${_REGION}-docker.pkg.dev/${PROJECT_ID}/${_ARTIFACT_REGISTRY_REPO}/${_SERVICE_NAME}:${COMMIT_SHA}",
+        "${_REGION}-docker.pkg.dev/${PROJECT_ID}/${_ARTIFACT_REGISTRY_REPO}/${_SERVICE_NAME}",
+        "--region",
+        "${_REGION}",
+        "--add-cloudsql-instances",
+        "${PROJECT_ID}:${_REGION}:${_INSTANCE_NAME}", 
+        "--platform", 
+        "managed",
+        "--allow-unauthenticated",   
+      ]
+
+substitutions:
+  _INSTANCE_NAME: test-instance
+  _REGION: northamerica-northeast1
+  # _CLOUD_RUN_SERVICE_NAME: polls-service
+  _SERVICE_NAME: "hello-world"
+  # _SERVICE_NAME: "polls-service"
+  _SECRET_SETTINGS_NAME: django_settings
+  _ARTIFACT_REGISTRY_REPO: "hello-world-app"
+  
+
+images:
+  # - "${_REGION}-docker.pkg.dev/${PROJECT_ID}/${_ARTIFACT_REGISTRY_REPO}/${_SERVICE_NAME}:${COMMIT_SHA}"
+  - "${_REGION}-docker.pkg.dev/${PROJECT_ID}/${_ARTIFACT_REGISTRY_REPO}/${_SERVICE_NAME}"
+
+# [END cloudrun_django_cloudbuild]
