--- djangoproject/entrypoint.sh
@@ -1,8 +1,0 @@
-#!/bin/sh
-
-# Apply database migrations
-python manage.py makemigrations
-python manage.py migrate
-
-# Start the server
-exec "$@"
