--- djangoproject/hello_world/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class HelloworldConfig(AppConfig):
-    default_auto_field = "django.db.models.BigAutoField"
-    name = "hello_world"
