--- djangoproject/requirements.txt
@@ -1,9 +1,0 @@
-asgiref==3.7.2
-black==23.3.0
-django==4.2.2
-psycopg2-binary==2.9.6
-pycodestyle==2.10.0
-gunicorn==20.1.0
-django-environ==0.10.0
-django-storages[google]==1.12.3
-google-cloud-secret-manager
