+++ hello_world/templates/delete_city.html
