+++ requirements.txt
@@ -1,0 +1,9 @@
+Django==4.2.2; python_version >= '3.8'
+Django==4.2.2; python_version < '3.8'
+django-storages[google]==1.13.2
+django-environ==0.10.0
+psycopg2-binary==2.9.6
+gunicorn==20.1.0
+google-cloud-secret-manager==2.16.1
+asgiref==3.7.2
+black==23.3.0
